requirements.txt
#with Python 3.12.3
# appdaemon for IDE when not in appdaemon
# having trouble with py 3.12, this should work
# pip install git+https://github.com/AppDaemon/appdaemon.git
#pip install appdaemon
hassapi==0.2.1
#pytapo
pytapo==3.3.38
# hassapiHASS plugin does not seem to be loading with numpy >= 2
numpy<2
#pip install python-ffmpeg
aiofiles==24.1.0
python-ffmpeg==2.0.12